// src/models/like.model.ts
import mongoose, { Schema, Document } from "mongoose";

export interface ILike extends Document {
    video: mongoose.Types.ObjectId;
    comment: mongoose.Types.ObjectId;
    tweet: mongoose.Types.ObjectId;
    likedBy: mongoose.Types.ObjectId;
    createdAt: Date;
    updatedAt: Date;
}

const likeSchema = new Schema<ILike>(
    {
        video: {
            type: Schema.Types.ObjectId,
            ref: "Video"
        },
        comment: {
            type: Schema.Types.ObjectId,
            ref: "Comment"
        },
        likedBy: {
            type: Schema.Types.ObjectId,
            ref: "User"
        },
    },
    {
        timestamps: true
    }
)

export const Like = mongoose.model<ILike>("Like", likeSchema);